Moved MultiDotInc to operators.py
